Move util to a more generic location
